move getObjectValueByPath
